Progetto finito

+++ negozio-online/backend/src/main/java/com/negozio/entita/Prodotto.java
@@ -1,41 +1,52 @@
 package com.negozio.entita;
 
 import javax.persistence.*;
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
+// questa e la classe che rappresenta i vini nel database
+// uso JPA per mappare automaticamente questa classe sulla tabella prodotti
 @Entity
 @Table(name = "prodotti")
 public class Prodotto {
     
+    // id del prodotto generato automaticamente dal database
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     
+    // nome del vino obbligatorio
     @Column(nullable = false)
     private String nome;
     
+    // descrizione del vino per i dettagli
     private String descrizione;
     
     @Column(nullable = false)
-    private BigDecimal prezzo;
+    private double prezzo;
     
+    // quante bottiglie abbiamo in magazzino
     private Integer quantita;
     
+    // link alla foto del vino
     private String urlImmagine;
     
+    // collegamento alla categoria rosso bianco rosato ecc
     @Column(name = "categoria_id")
     private Long categoriaId;
     
+    // per nascondere prodotti senza cancellarli dal db
     private Boolean attivo = true;
     
+    // quando ho aggiunto questo vino al catalogo
     private LocalDateTime dataCreazione;
 
+    // costruttore vuoto richiesto da JPA
     public Prodotto() {
         this.dataCreazione = LocalDateTime.now();
     }
     
-    public Prodotto(String nome, String descrizione, BigDecimal prezzo) {
+    // costruttore per creare un nuovo prodotto
+    public Prodotto(String nome, String descrizione, double prezzo) {
         this.nome = nome;
         this.descrizione = descrizione;
         this.prezzo = prezzo;
@@ -43,6 +54,8 @@ public class Prodotto {
         this.attivo = true;
     }
 
+    // metodi getter e setter per accedere ai campi privati
+    
     public Long getId() {
         return id;
     }
@@ -67,11 +80,11 @@ public class Prodotto {
         this.descrizione = descrizione;
     }
 
-    public BigDecimal getPrezzo() {
+    public double getPrezzo() {
         return prezzo;
     }
 
-    public void setPrezzo(BigDecimal prezzo) {
+    public void setPrezzo(double prezzo) {
         this.prezzo = prezzo;
     }
